Fix issue with get costume value block
NGL, SharkPool totally blackmailed me into fixing this bug. Okay,
legally he didn't, but he told me to fix a bug that I didn't cause and
was in an entirely different part of the codebase before he would merge
the other patch I made.

Just for full clarity, that patch is

<https://github.com/PenguinMod/penguinmod.github.io/pull/97>

Which is in the GUI, fixing this bug required a fix to the VM (here).

Someone else really should've spotted and fixed this before, but
whatever, guess I've gotta be the one to fix it because otherwise a very
good change (above) isn't getting merged.

IDEK why I've made this commit message so long, I'm not even gonna be
using this as the PR message.