Add support for TurboWarp's extensionStorage key
TurboWarp uses the key extensionStorage rather than the one PenguinMod
uses: "extensionData." This discrepancy means that projects made in
TurboWarp that use extensions that use the extensionStorage API cannot
be imported into PenguinMod without potentially losing crucial data.

I would recommend against switching over to extensionStorage as the key
in future commits, as it would require an update to the protobufs lib,
which is annoying, and I'm not gonna be the one to deal with that mess
of code.